# api/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "contact": {}
  },
  "paths": {
...
    "/user": {
      "delete": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "The account will be soft deleted.\nIf the user is the last admin, the account cannot be deleted.\nIf a user is soft-deleted, the account will be permanently deleted in the future.",
        "produces": ["application/json"],
        "tags": ["User"],
        "summary": "Delete user account",
        "responses": {
          "200": {
            "description": "Password reset",
            "schema": {
              "$ref": "#/definitions/JSendSuccess-map_string_string"
            }
          },
          "400": {
            "description": "Last admin account",
            "schema": {
              "$ref": "#/definitions/JSendFailure-ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/JSendFailure-ErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/JSendError"
            }
          }
        }
      }
    },
    "/user/email": {
      "patch": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Requests an email change for the user. An email change token will be sent to the new email address.",
        "produces": ["application/json"],
        "tags": ["User"],
        "summary": "Request email change for the user",
        "parameters": [
          {
            "description": "Request body",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ChangeEmailQuery"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Email change request token sent",
            "schema": {
              "$ref": "#/definitions/JSendSuccess-map_string_string"
            }
          },
          "400": {
            "description": "Bad request/invalid email/already in use",
            "schema": {
              "$ref": "#/definitions/JSendFailure-ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/JSendFailure-ErrorResponse"
            }
          },
          "422": {
            "description": "Bad query format",
            "schema": {
              "$ref": "#/definitions/JSendFailure-ValidationResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/JSendError"
            }
          }
        }
      }
    },
    "/user/email/confirm": {
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Confirms an email change for the user.\nThe new email address will be active on the next login.\nYou have to login (authenticate) with the old email address to confirm the change.",
        "produces": ["application/json"],
        "tags": ["User"],
        "summary": "Confirm email change for the user",
        "parameters": [
          {
            "description": "Request body",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ConfirmEmailChangeQuery"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Email changed",
            "schema": {
              "$ref": "#/definitions/JSendSuccess-map_string_string"
            }
          },
          "400": {
            "description": "Invalid token",
            "schema": {
              "$ref": "#/definitions/JSendFailure-ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/JSendFailure-ErrorResponse"
            }
          },
          "403": {
            "description": "Token expired",
            "schema": {
              "$ref": "#/definitions/JSendFailure-ErrorResponse"
            }
          },
          "404": {
            "description": "No email change request found",
            "schema": {
              "$ref": "#/definitions/JSendFailure-ErrorResponse"
            }
          },
          "422": {
            "description": "Bad query format",
            "schema": {
              "$ref": "#/definitions/JSendFailure-ValidationResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/JSendError"
            }
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "description": "Acciqures a JWT token for the user to access the API\nOnly active users can login\nUsers can downgrade their role by providing the role in the request (optional)",
        "produces": ["application/json"],
        "tags": ["Login"],
        "summary": "Login for the API to get JWT token",
        "parameters": [
          {
            "description": "Request body",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LoginQuery"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JWT token",
            "schema": {
              "$ref": "#/definitions/JSendSuccess-LoginResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/JSendFailure-ErrorResponse"
            }
          },
          "403": {
            "description": "User is not active",
            "schema": {
              "$ref": "#/definitions/JSendFailure-ErrorResponse"
            }
          },
          "422": {
            "description": "Bad query format",
            "schema": {
              "$ref": "#/definitions/JSendFailure-ValidationResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/JSendError"
            }
          }
        }
      }
    },
    "/user/password": {
      "patch": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Changes the password for the user. The old password must be provided.\nThe new password will be active on the next login.",
        "produces": ["application/json"],
        "tags": ["User"],
        "summary": "Change password for the user",
        "parameters": [
          {
            "description": "Request body",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ChangePwdQuery"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Password changed",
            "schema": {
              "$ref": "#/definitions/JSendSuccess-map_string_string"
            }
          },
          "400": {
            "description": "Wrong old password/invalid new password",
            "schema": {
              "$ref": "#/definitions/JSendFailure-ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/JSendFailure-ErrorResponse"
            }
          },
          "422": {
            "description": "Bad query format",
            "schema": {
              "$ref": "#/definitions/JSendFailure-ValidationResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/JSendError"
            }
          }
        }
      }
    },
    "/user/password/reset": {
      "post": {
        "description": "Requests a password reset for the user. A password reset token will be sent to the user's email.\nPassword reset tokens are valid for a limited time.\nThe API will always return the same message (200) to prevent email enumeration.",
        "produces": ["application/json"],
        "tags": ["Login"],
        "summary": "Request password reset",
        "parameters": [
          {
            "description": "Request body",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ResetPwdQuery"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Password reset token sent",
            "schema": {
              "$ref": "#/definitions/JSendSuccess-map_string_string"
            }
          },
          "422": {
            "description": "Bad query format",
            "schema": {
              "$ref": "#/definitions/JSendFailure-ValidationResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/JSendError"
            }
          }
        }
      }
    },
    "/user/password/reset/confirm": {
      "post": {
        "description": "Confirms a password reset or first password set for the user.\nThe API will always return the same message (400) on auth errors to prevent email enumeration.",
        "produces": ["application/json"],
        "tags": ["Login"],
        "summary": "Confirm password reset or first password set",
        "parameters": [
          {
            "description": "Request body",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ResetConfirmPwdQuery"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Password reset",
            "schema": {
              "$ref": "#/definitions/JSendSuccess-map_string_string"
            }
          },
          "400": {
            "description": "Bad request/invalid token",
            "schema": {
              "$ref": "#/definitions/JSendFailure-ErrorResponse"
            }
          },
          "403": {
            "description": "Token expired",
            "schema": {
              "$ref": "#/definitions/JSendFailure-ErrorResponse"
            }
          },
          "422": {
            "description": "Bad query format",
            "schema": {
              "$ref": "#/definitions/JSendFailure-ValidationResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/JSendError"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "ChangeEmailQuery": {
      "type": "object",
      "required": ["email"],
      "properties": {
        "email": {
          "description": "New email address",
          "type": "string",
          "example": "newmail@newcomp.com"
        }
      }
    },
...
    "JSendError": {
      "type": "object",
      "properties": {
        "message": {
          "description": "Error message",
          "type": "string",
          "example": "Internal server error"
        },
        "status": {
          "description": "Status",
          "type": "string",
          "example": "error"
        }
      }
    },
    "JSendFailure-ErrorResponse": {
      "type": "object",
      "properties": {
        "data": {
          "description": "Data with error message(s)",
          "allOf": [
            {
              "$ref": "#/definitions/ErrorResponse"
            }
          ]
        },
        "status": {
          "description": "Status 'fail'",
          "type": "string",
          "example": "fail"
        }
      }
    },
    "JSendFailure-ValidationResponse": {
      "type": "object",
      "properties": {
        "data": {
          "description": "Data with error message(s)",
          "allOf": [
            {
              "$ref": "#/definitions/ValidationResponse"
            }
          ]
        },
        "status": {
          "description": "Status 'fail'",
          "type": "string",
          "example": "fail"
        }
      }
    },
...
    "JSendSuccess-map_string_string": {
      "type": "object",
      "properties": {
        "data": {
          "description": "Data with success message(s)",
          "allOf": [
            {
              "$ref": "#/definitions/map_string_string"
            }
          ]
        },
        "status": {
          "description": "Status 'success'",
          "type": "string",
          "example": "success"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "description": "Type \"Bearer\" followed by a space and JWT token.",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}